refactor: Separate large file into several files
