+++ windows/main.cpp
@@ -1,13 +1,4 @@
-#define WINVER 0x0605
-#include <windows.h>
-#include <windowsx.h>
-#include <commctrl.h>
-#include <d3d11.h>
-#include <d3d11_2.h>
-#include <dcomp.h>
-#include <dwmapi.h>
-#include <dxgi.h>
-#include <wrl.h>
+#include "main.hpp"
 #include <memory>
 #include <string>
 #include <string_view>
@@ -17,187 +8,10 @@
 #include "../constant.hpp"
 #include "../keyboard.hpp"
 #include "../platform_api.hpp"
-#include "../util.hpp"
-#include "../viewer.hpp"
 #include "glm/vec2.hpp"  // IWYU pragma: keep; silence clangd.
-#include "sokol_gfx.h"
+#include "menu.hpp"
+#include "msgbox.hpp"
 #include "sokol_time.h"
-
-template <typename T>
-using ComPtr = Microsoft::WRL::ComPtr<T>;
-
-namespace {
-SIZE rectToSize(RECT rect);
-std::vector<HMONITOR> getAllMonitorHandles();
-std::optional<HMONITOR> getMonitorHandleFromID(int monitorID);
-std::optional<RECT> getMonitorWorkareaFromID(int monitorID);
-
-template <typename T, typename DeleteFunc, DeleteFunc deleteFunc>
-class UniqueHandler {
-public:
-    UniqueHandler(T handler) : handler_(handler, deleteFunc) {};
-    UniqueHandler() : UniqueHandler(nullptr) {};
-    T GetRawHandler() { return handler_.get(); };
-    operator T() { return GetRawHandler(); };
-    UniqueHandler& operator=(T& handler) {
-        handler_.reset(handler);
-        return *this;
-    };
-
-private:
-    std::unique_ptr<std::remove_pointer_t<T>, DeleteFunc> handler_;
-};
-
-using UniqueHWND = UniqueHandler<HWND, decltype(&DestroyWindow), &DestroyWindow>;
-using UniqueHMENU = UniqueHandler<HMENU, decltype(&DestroyMenu), &DestroyMenu>;
-}  // namespace
-
-class AppMenu {
-public:
-    AppMenu();
-    ~AppMenu();
-    void Setup();
-    void Terminate();
-    void ShowMenu();
-    bool IsMenuOpened() const;
-
-public:
-    static constexpr UINT YOMMD_WM_TOGGLE_ENABLE_MOUSE = WM_APP;
-    static constexpr UINT YOMMD_WM_SHOW_TASKBAR_MENU = WM_APP + 1;
-
-private:
-    static DWORD WINAPI showMenu(LPVOID param);
-    static LRESULT CALLBACK windowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
-
-    void createTaskbar();
-
-private:
-    class Cmd {
-    private:
-        template <typename LHS, typename RHS>
-        using SelectSmallerSizeType = std::conditional<sizeof(LHS) < sizeof(RHS), LHS, RHS>;
-
-        template <typename HD, typename... TL>
-        struct MinimumSizeType {
-            using type =
-                typename SelectSmallerSizeType<HD, typename MinimumSizeType<TL...>::type>::
-                    type;
-        };
-
-        template <typename T>
-        struct MinimumSizeType<T> {
-            using type = T;
-        };
-
-    public:
-        using UnderlyingType = MinimumSizeType<UINT_PTR, UINT, WORD>::type;
-        enum class Kind : UnderlyingType {
-            None,
-            EnableMouse,
-            ResetPosition,
-            SelectScreen,
-            HideWindow,
-            Quit,
-            MenuCount,
-        };
-        using enum Kind;
-
-        static constexpr Kind GetCmd(UnderlyingType cmd);
-        static constexpr UnderlyingType GetUserData(UnderlyingType cmd);
-        static constexpr UnderlyingType Combine(Kind cmd, UnderlyingType userData);
-
-    private:
-        static constexpr size_t fieldLength_ = sizeof(UnderlyingType) * 8 / 2;
-
-        static_assert(
-            Enum::underlyCast(Kind::MenuCount) < (UnderlyingType(1) << fieldLength_),
-            "Too many menu commands declared");
-    };
-
-    static constexpr PCWSTR wcMenuName = L"yoMMD-menu-window";
-    static constexpr PCWSTR wcSelectorName = L"yoMMD-screen-selector-window";
-
-    HANDLE hMenuThread_;
-    HICON hTaskbarIcon_;
-    NOTIFYICONDATAW taskbarIconDesc_;
-    bool isMenuOpened_;
-};
-
-class AppMain {
-public:
-    AppMain();
-    ~AppMain();
-    void Setup(const CmdArgs& cmdArgs);
-    void UpdateDisplay();
-    void Terminate();
-    bool IsRunning() const;
-    void ChangeScreen(int screenID);
-    Routine& GetRoutine();
-    const HWND& GetWindowHandle() const;
-    sg_environment GetSokolEnvironment() const;
-    sg_swapchain GetSokolSwapchain() const;
-    glm::vec2 GetWindowSize() const;
-    glm::vec2 GetDrawableSize() const;
-    int GetSampleCount() const;
-    bool IsMenuOpened() const;
-
-private:
-    static LRESULT CALLBACK windowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
-
-    void createWindow();
-    void createDrawable();
-    int determineSampleCount() const;
-    LRESULT handleMessage(UINT uMsg, WPARAM wParam, LPARAM lParam);
-
-private:
-    static constexpr PCWSTR windowClassName_ = L"yoMMD AppMain";
-
-    bool isRunning_;
-    int sampleCount_;
-    Routine routine_;
-    AppMenu menu_;
-    HWND hwnd_;
-    ComPtr<IDXGISwapChain1> swapChain_;
-    ComPtr<ID3D11Texture2D> renderTarget_;
-    ComPtr<ID3D11RenderTargetView> renderTargetView_;
-    ComPtr<ID3D11Texture2D> msaaRenderTarget_;
-    ComPtr<ID3D11RenderTargetView> msaaRenderTargetView_;
-    ComPtr<ID3D11Device> d3Device_;
-    ComPtr<ID3D11DeviceContext> deviceContext_;
-    ComPtr<IDXGIDevice> dxgiDevice_;
-    ComPtr<IDXGIFactory2> dxFactory_;
-    ComPtr<ID3D11Texture2D> depthStencilBuffer_;
-    ComPtr<ID3D11DepthStencilView> depthStencilView_;
-    ComPtr<IDCompositionDevice> dcompDevice_;
-    ComPtr<IDCompositionTarget> dcompTarget_;
-    ComPtr<IDCompositionVisual> dcompVisual_;
-};
-
-class MsgBox {
-public:
-    static void Init();
-    static void Terminate();
-    static void Show(std::string_view);
-
-private:
-    static LRESULT CALLBACK windowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam);
-    static void drawContents(HWND hwnd);
-
-private:
-    constexpr static PCWSTR className_ = L"yoMMD-messagebox";
-    constexpr static int okMenuID_ = 100;
-    constexpr static UINT winStyle_ = WS_CAPTION | WS_SYSMENU | WS_TABSTOP | WS_DLGFRAME;
-    static bool initialized_;
-    static bool showingWindow_;
-    static HINSTANCE hInstance_;
-    static HFONT hfont_;
-    static HWND buttonHWND_;
-    static std::wstring wmsg_;
-};
-
-namespace {
-inline AppMain& getAppMain();
-}  // namespace
 
 AppMain::AppMain() :
     isRunning_(true), sampleCount_(Constant::PreferredSampleCount), hwnd_(nullptr) {}
@@ -597,431 +411,6 @@ LRESULT AppMain::handleMessage(UINT uMsg, WPARAM wParam, LPARAM lParam) {
     return DefWindowProcW(hwnd_, uMsg, wParam, lParam);
 }
 
-AppMenu::AppMenu() : hMenuThread_(nullptr), hTaskbarIcon_(nullptr) {}
-
-AppMenu::~AppMenu() {
-    Terminate();
-}
-
-void AppMenu::Setup() {
-    WNDCLASSW wc = {};
-
-    wc.style = CS_HREDRAW | CS_VREDRAW | CS_OWNDC;
-    wc.lpfnWndProc = AppMenu::windowProc, wc.hInstance = GetModuleHandleW(nullptr);
-    wc.lpszClassName = wcMenuName;
-    wc.hIcon = LoadIcon(nullptr, IDI_WINLOGO);
-    wc.hCursor = LoadCursor(nullptr, IDC_ARROW);
-    RegisterClassW(&wc);
-
-    wc.lpfnWndProc = DefWindowProcW;
-    wc.lpszClassName = wcSelectorName;
-    wc.hbrBackground = reinterpret_cast<HBRUSH>(GetStockObject(BLACK_BRUSH));
-    RegisterClassW(&wc);
-
-    createTaskbar();
-}
-
-void AppMenu::Terminate() {
-    if (hTaskbarIcon_) {
-        DestroyIcon(hTaskbarIcon_);
-        hTaskbarIcon_ = nullptr;
-    }
-
-    Shell_NotifyIconW(NIM_DELETE, &taskbarIconDesc_);
-    UnregisterClassW(wcMenuName, GetModuleHandleW(nullptr));
-
-    DWORD exitCode;
-    if (GetExitCodeThread(hMenuThread_, &exitCode) && exitCode == STILL_ACTIVE) {
-        Info::Log("Thread is still running. Wait finishing.");
-        // Right click menu/toolbar menu must not appear so long.
-        // We should be able to wait for it is closed.
-        WaitForSingleObject(hMenuThread_, INFINITE);
-        CloseHandle(hMenuThread_);
-    }
-}
-
-void AppMenu::ShowMenu() {
-    DWORD exitCode;
-    if (GetExitCodeThread(hMenuThread_, &exitCode) && exitCode == STILL_ACTIVE) {
-        Info::Log("Thread is running");
-    }
-
-    hMenuThread_ = CreateThread(NULL, 0, AppMenu::showMenu, this, 0, NULL);
-}
-
-bool AppMenu::IsMenuOpened() const {
-    return isMenuOpened_;
-}
-
-DWORD WINAPI AppMenu::showMenu(LPVOID param) {
-    AppMenu *appMenu = reinterpret_cast<AppMenu *>(param);
-    constexpr DWORD winStyle = WS_CHILD;
-    const HWND& parentWin = getAppMain().GetWindowHandle();
-    UniqueHWND hMenuWindow, hSelectorWindow;
-
-    const LONG parentWinExStyle = GetWindowLongW(parentWin, GWL_EXSTYLE);
-    if (parentWinExStyle == 0) {
-        Info::Log("Failed to get parent window's style");
-    }
-
-    hSelectorWindow = CreateWindowExW(
-        WS_EX_LAYERED | WS_EX_NOACTIVATE, wcSelectorName, L"", WS_DISABLED | WS_POPUP, 0, 0, 0,
-        0, nullptr, nullptr, GetModuleHandleW(nullptr), nullptr);
-    if (!hSelectorWindow.GetRawHandler()) {
-        Err::Log("Failed to create screen selector window.");
-        return 1;
-    }
-    SetLayeredWindowAttributes(hSelectorWindow, RGB(0, 0, 0), 127, LWA_ALPHA);
-    BOOL fDisable = TRUE;
-    DwmSetWindowAttribute(
-        hSelectorWindow, DWMWA_TRANSITIONS_FORCEDISABLED, &fDisable, sizeof(fDisable));
-
-    hMenuWindow = CreateWindowExW(
-        0, wcMenuName, L"", winStyle, 0, 0, 0, 0, parentWin, nullptr,
-        GetModuleHandleW(nullptr), hSelectorWindow);
-    if (!hMenuWindow.GetRawHandler()) {
-        Err::Log("Failed to create dummy window for menu.");
-        return 1;
-    }
-
-    POINT point;
-    if (!GetCursorPos(&point)) {
-        Err::Log("Failed to get mouse point");
-        return 1;
-    }
-
-    UniqueHMENU hScreensMenu = CreatePopupMenu();
-    const HMONITOR curMonitorHandle =
-        MonitorFromWindow(getAppMain().GetWindowHandle(), MONITOR_DEFAULTTONULL);
-    std::vector<HMONITOR> monitorHandles = getAllMonitorHandles();
-    for (int cnt = monitorHandles.size(), i = 0; i < cnt; ++i) {
-        const std::wstring title(L"&Screen" + std::to_wstring(i));
-        const auto op = Cmd::Combine(Cmd::SelectScreen, i);
-        AppendMenuW(hScreensMenu, MF_STRING, op, title.c_str());
-        if (monitorHandles[i] == curMonitorHandle) {
-            EnableMenuItem(hScreensMenu, op, MF_DISABLED);
-        }
-    }
-
-    UniqueHMENU hmenu = CreatePopupMenu();
-    if (parentWinExStyle & WS_EX_TRANSPARENT) {
-        AppendMenuW(hmenu, MF_STRING, Enum::underlyCast(Cmd::EnableMouse), L"&Enable Mouse");
-    } else {
-        AppendMenuW(hmenu, MF_STRING, Enum::underlyCast(Cmd::EnableMouse), L"&Disable Mouse");
-    }
-    AppendMenuW(hmenu, MF_STRING, Enum::underlyCast(Cmd::ResetPosition), L"&Reset Position");
-    AppendMenuW(hmenu, MF_SEPARATOR, Enum::underlyCast(Cmd::None), L"");
-    AppendMenuW(
-        hmenu, MF_POPUP, reinterpret_cast<UINT_PTR>(hScreensMenu.GetRawHandler()),
-        L"&Select screen");
-    AppendMenuW(hmenu, MF_SEPARATOR, Enum::underlyCast(Cmd::None), L"");
-    if (GetWindowLongPtrW(parentWin, GWL_STYLE) & WS_VISIBLE)
-        AppendMenuW(hmenu, MF_STRING, Enum::underlyCast(Cmd::HideWindow), L"&Hide Window");
-    else
-        AppendMenuW(hmenu, MF_STRING, Enum::underlyCast(Cmd::HideWindow), L"&Show Window");
-    AppendMenuW(hmenu, MF_SEPARATOR, Enum::underlyCast(Cmd::None), L"");
-    AppendMenuW(hmenu, MF_STRING, Enum::underlyCast(Cmd::Quit), L"&Quit");
-
-    if (parentWinExStyle == 0) {
-        EnableMenuItem(hmenu, Enum::underlyCast(Cmd::EnableMouse), MF_DISABLED);
-    }
-
-    if (monitorHandles.size() <= 1)
-        EnableMenuItem(
-            hmenu, reinterpret_cast<UINT_PTR>(hScreensMenu.GetRawHandler()), MF_DISABLED);
-
-    constexpr UINT menuFlags = TPM_RIGHTBUTTON | TPM_NONOTIFY | TPM_RETURNCMD;
-
-    SetForegroundWindow(hMenuWindow);
-    appMenu->isMenuOpened_ = true;
-    const auto op = TrackPopupMenuEx(hmenu, menuFlags, point.x, point.y, hMenuWindow, nullptr);
-    appMenu->isMenuOpened_ = false;
-
-    switch (Cmd::GetCmd(op)) {
-    case Cmd::EnableMouse:
-        if (parentWinExStyle != 0) {
-            SetWindowLongW(parentWin, GWL_EXSTYLE, parentWinExStyle ^ WS_EX_TRANSPARENT);
-        }
-        break;
-    case Cmd::ResetPosition:
-        getAppMain().GetRoutine().ResetModelPosition();
-        break;
-    case Cmd::SelectScreen:
-        getAppMain().ChangeScreen(Cmd::GetUserData(op));
-        break;
-    case Cmd::HideWindow:
-        if (GetWindowLongPtrW(parentWin, GWL_STYLE) & WS_VISIBLE)
-            ShowWindow(parentWin, SW_HIDE);
-        else
-            ShowWindow(parentWin, SW_SHOWNORMAL);
-        break;
-    case Cmd::Quit:
-        SendMessageW(parentWin, WM_DESTROY, 0, 0);
-        break;
-    case Cmd::None:
-        // Menu is canceled.
-        // The mouse click to cancel right-click menu will dispatch
-        // WM_MOUSEMOVE message and it leads wrong and fake mouse drag event.
-        // As a workaround, in order to avoid this call OnGestureEnd to update
-        // the action context.
-        getAppMain().GetRoutine().OnGestureEnd();
-        break;
-    case Cmd::MenuCount:
-        Err::Log("Internal error: Command::MenuCount is used");
-        break;
-    }
-
-    return 0;
-}
-
-LRESULT CALLBACK AppMenu::windowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam) {
-    switch (uMsg) {
-    case WM_MENUSELECT:
-        if (HIWORD(wParam) & MF_MOUSESELECT && !(HIWORD(wParam) & MF_POPUP)) {
-            const HWND hSelectorWindow =
-                reinterpret_cast<HWND>(GetWindowLongPtrW(hwnd, GWLP_USERDATA));
-            ShowWindow(hSelectorWindow, SW_HIDE);
-
-            const HMENU hmenu = reinterpret_cast<HMENU>(lParam);
-            const std::underlying_type_t<Cmd::Kind> op = LOWORD(wParam);
-            MENUITEMINFOW itemInfo = {
-                .cbSize = sizeof(itemInfo),
-                .fMask = MIIM_FTYPE | MIIM_STATE | MIIM_ID,
-            };
-            GetMenuItemInfoW(hmenu, op, false, &itemInfo);
-            if (itemInfo.fType != MFT_STRING || itemInfo.fState & MFS_DISABLED)
-                break;
-            if (Cmd::GetCmd(op) != Cmd::SelectScreen)
-                break;
-
-            const auto r = getMonitorWorkareaFromID(Cmd::GetUserData(op));
-            if (!r.has_value()) {
-                // Maybe monitor is disconnected. Do nothing.
-                break;
-            }
-
-            const auto cx = r->right - r->left;
-            const auto cy = r->bottom - r->top;
-            constexpr UINT uFlags = SWP_SHOWWINDOW | SWP_NOACTIVATE;
-            SetWindowPos(hSelectorWindow, HWND_TOPMOST, r->left, r->top, cx, cy, uFlags);
-        }
-        break;
-    case WM_CREATE: {
-        const CREATESTRUCT& cs = *reinterpret_cast<CREATESTRUCT *>(lParam);
-        SetWindowLongPtrW(hwnd, GWLP_USERDATA, reinterpret_cast<LONG_PTR>(cs.lpCreateParams));
-    } break;
-    }
-    return DefWindowProcW(hwnd, uMsg, wParam, lParam);
-}
-
-void AppMenu::createTaskbar() {
-    const HWND& parentWin = getAppMain().GetWindowHandle();
-    if (!parentWin) {
-        Err::Exit(
-            "Internal error:", "Taskbar must be created after the main window is created:",
-            __FILE__, __LINE__);
-    }
-
-    auto iconData = Resource::getStatusIconData();
-    hTaskbarIcon_ = CreateIconFromResource(
-        const_cast<PBYTE>(iconData.data()), iconData.length(), TRUE, 0x00030000);
-    if (!hTaskbarIcon_) {
-        Err::Log("Failed to load icon. Fallback to Windows' default application icon.");
-        hTaskbarIcon_ = LoadIconA(nullptr, IDI_APPLICATION);
-        if (!hTaskbarIcon_) {
-            Err::Exit("Icon fallback failed.");
-        }
-    }
-
-    taskbarIconDesc_.cbSize = sizeof(taskbarIconDesc_);
-    taskbarIconDesc_.hWnd = parentWin;
-    taskbarIconDesc_.uID = 100;  // TODO: What value should be here?
-    taskbarIconDesc_.hIcon = hTaskbarIcon_;
-    taskbarIconDesc_.uVersion = NOTIFYICON_VERSION_4;
-    taskbarIconDesc_.uCallbackMessage = YOMMD_WM_SHOW_TASKBAR_MENU;
-    wcscpy_s(taskbarIconDesc_.szTip, sizeof(taskbarIconDesc_.szTip), L"yoMMD");
-    taskbarIconDesc_.uFlags = NIF_ICON | NIF_TIP | NIF_SHOWTIP | NIF_MESSAGE;
-
-    Shell_NotifyIconW(NIM_ADD, &taskbarIconDesc_);
-}
-
-constexpr AppMenu::Cmd::Kind AppMenu::Cmd::GetCmd(AppMenu::Cmd::UnderlyingType cmd) {
-    constexpr UnderlyingType mask = (UnderlyingType(1) << fieldLength_) - 1;
-    return Kind(cmd & mask);
-}
-
-constexpr AppMenu::Cmd::UnderlyingType AppMenu::Cmd::GetUserData(
-    AppMenu::Cmd::UnderlyingType cmd) {
-    constexpr UnderlyingType mask = (UnderlyingType(1) << fieldLength_) - 1;
-    return ((cmd >> fieldLength_) & mask);
-}
-
-constexpr AppMenu::Cmd::UnderlyingType AppMenu::Cmd::Combine(
-    AppMenu::Cmd::Kind kind,
-    AppMenu::Cmd::UnderlyingType userData) {
-    userData <<= fieldLength_;
-    return Enum::underlyCast(kind) | userData;
-}
-
-bool MsgBox::initialized_ = false;
-bool MsgBox::showingWindow_ = false;
-HINSTANCE MsgBox::hInstance_ = nullptr;
-HFONT MsgBox::hfont_ = nullptr;
-HWND MsgBox::buttonHWND_ = nullptr;
-std::wstring MsgBox::wmsg_{};
-
-void MsgBox::Init() {
-    hInstance_ = GetModuleHandleW(nullptr);
-    hfont_ = static_cast<HFONT>(GetStockObject(OEM_FIXED_FONT));
-
-    WNDCLASSW wc = {};
-    wc.style = CS_HREDRAW | CS_VREDRAW;
-    wc.lpfnWndProc = windowProc;
-    wc.cbClsExtra = 0;
-    wc.cbWndExtra = 0;
-    wc.hInstance = hInstance_;
-    wc.hIcon = LoadIcon(nullptr, IDI_WARNING);
-    wc.hCursor = LoadCursor(nullptr, IDC_ARROW);
-    wc.hbrBackground = reinterpret_cast<HBRUSH>(COLOR_WINDOW + 1);
-    wc.lpszMenuName = nullptr;
-    wc.lpszClassName = className_;
-
-    RegisterClassW(&wc);
-
-    initialized_ = true;
-}
-
-void MsgBox::Terminate() {
-    UnregisterClassW(className_, hInstance_);
-    wmsg_.clear();
-    hInstance_ = nullptr;
-    initialized_ = false;
-}
-
-void MsgBox::Show(std::string_view msg) {
-    if (!initialized_ || showingWindow_)
-        return;
-
-    const int size = MultiByteToWideChar(CP_UTF8, MB_COMPOSITE, msg.data(), -1, nullptr, 0);
-    wmsg_.resize(size - 1, '\0');
-    MultiByteToWideChar(CP_UTF8, MB_COMPOSITE, msg.data(), -1, wmsg_.data(), size);
-
-    const HWND hwnd = CreateWindowW(
-        className_, L"yoMMD Error", winStyle_, CW_USEDEFAULT, CW_USEDEFAULT, CW_USEDEFAULT,
-        CW_USEDEFAULT, nullptr, nullptr, hInstance_, nullptr);
-
-    if (!hwnd)
-        return;
-
-    ShowWindow(hwnd, SW_SHOWDEFAULT);
-    SetForegroundWindow(hwnd);
-
-    showingWindow_ = true;
-
-    MSG procMsg;
-    while (showingWindow_ && GetMessageW(&procMsg, nullptr, 0, 0)) {
-        if (!IsDialogMessageW(hwnd, &procMsg)) {
-            TranslateMessage(&procMsg);
-            DispatchMessage(&procMsg);
-        }
-    }
-}
-
-LRESULT CALLBACK MsgBox::windowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam) {
-    switch (uMsg) {
-    case WM_CREATE: {
-        constexpr UINT winStyle =
-            WS_CHILD | WS_VISIBLE | BS_CENTER | BS_VCENTER | BS_DEFPUSHBUTTON;
-
-        buttonHWND_ = CreateWindowW(
-            L"BUTTON", L"OK", winStyle, 0, 0, 60, 25, hwnd, reinterpret_cast<HMENU>(okMenuID_),
-            hInstance_, nullptr);
-        // FIXME: Button is not selected in default when showing two
-        // error dialog continuously.
-        SetForegroundWindow(buttonHWND_);
-        return 0;
-    }
-    case WM_PAINT:
-        drawContents(hwnd);
-        return 0;
-    case WM_COMMAND:
-        if (LOWORD(wParam) == okMenuID_) {
-            DestroyWindow(hwnd);
-            return 0;
-        }
-        break;
-    case WM_DESTROY:
-        DestroyWindow(buttonHWND_);
-        wmsg_.clear();
-        showingWindow_ = false;
-        buttonHWND_ = nullptr;
-        return 0;
-    }
-    return DefWindowProcW(hwnd, uMsg, wParam, lParam);
-}
-
-void MsgBox::drawContents(HWND hwnd) {
-    constexpr int textMerginX = 50;
-    constexpr int textMerginY = 25;
-    constexpr int buttonMerginY = 5;
-
-    HDC hdc;
-    HFONT hPrevFont;
-    PAINTSTRUCT ps = {};
-
-    RECT bounds = {};
-
-    RECT buttonRect = {};
-    SIZE buttonPos = {};
-    SIZE buttonSize = {};
-
-    SIZE screenSize = {};
-
-    GetClientRect(buttonHWND_, &buttonRect);
-    buttonSize = rectToSize(buttonRect);
-
-    hdc = BeginPaint(hwnd, &ps);
-
-    // Change font.
-    hPrevFont = static_cast<HFONT>(SelectObject(hdc, hfont_));
-
-    DrawTextW(hdc, wmsg_.data(), -1, &bounds, DT_CALCRECT);
-
-    // Calculate window content area.
-    RECT contentRect(bounds);
-    contentRect.right += textMerginX * 2;
-    contentRect.bottom += textMerginY * 2 + buttonSize.cy + buttonMerginY;
-
-    // Adjust window position.
-    // TODO: Check if window size over screen size.  If window is bigger than
-    // screen, we need to make font and then window make small.
-    RECT winRect(contentRect);
-    screenSize.cx = GetSystemMetrics(SM_CXSCREEN);
-    screenSize.cy = GetSystemMetrics(SM_CYSCREEN);
-    AdjustWindowRect(&winRect, winStyle_, TRUE);
-    SIZE winSize(rectToSize(winRect));
-    winRect.left = (screenSize.cx - winSize.cx) / 2;
-    winRect.top = (screenSize.cy - winSize.cy) / 2;
-    MoveWindow(hwnd, winRect.left, winRect.top, winSize.cx, winSize.cy, FALSE);
-
-    // Adjust button position.
-    buttonPos = rectToSize(contentRect);
-    buttonPos.cx -= textMerginX + buttonSize.cx;
-    buttonPos.cy -= buttonMerginY + buttonSize.cy;
-    MoveWindow(buttonHWND_, buttonPos.cx, buttonPos.cy, buttonSize.cx, buttonSize.cy, FALSE);
-
-    // Adjust text position.
-    bounds.right = bounds.right - bounds.left + textMerginX * 2;
-    bounds.bottom = bounds.bottom - bounds.top + textMerginY * 2;
-    bounds.left = textMerginX;
-    bounds.top = textMerginY;
-    DrawTextW(hdc, wmsg_.data(), -1, &bounds, DT_LEFT);
-
-    // Restore font.
-    SelectObject(hdc, hPrevFont);
-    EndPaint(hwnd, &ps);
-}
-
 namespace Context {
 sg_environment getSokolEnvironment() {
     return getAppMain().GetSokolEnvironment();
@@ -1075,10 +464,6 @@ void messageBox(std::string_view msg) {
 }
 }  // namespace Dialog
 
-namespace {
-SIZE rectToSize(RECT rect) {
-    return {rect.right - rect.left, rect.bottom - rect.top};
-}
 std::vector<HMONITOR> getAllMonitorHandles() {
     static const MONITORENUMPROC proc = [](HMONITOR hMonitor, HDC hdc, LPRECT rect,
                                            LPARAM param) -> BOOL {
@@ -1127,8 +512,6 @@ AppMain& getAppMain() {
     static AppMain appMain{};
     return appMain;
 }
-
-}  // namespace
 
 int WINAPI wWinMain(HINSTANCE hInstance, HINSTANCE, PWSTR pCmdLine, int nCmdShow) {
     (void)hInstance;
